refactor: command/query 패키지로 정리
