+++ src/main/java/umc/cockple/demo/domain/chat/presentation/rest/ChatController.java
@@ -9,9 +9,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import umc.cockple.demo.domain.chat.dto.*;
-import umc.cockple.demo.domain.chat.service.ChatCommandService;
+import umc.cockple.demo.domain.chat.service.command.ChatCommandService;
 import umc.cockple.demo.domain.chat.service.ChatFileService;
-import umc.cockple.demo.domain.chat.service.ChatQueryService;
+import umc.cockple.demo.domain.chat.service.query.ChatQueryService;
 import umc.cockple.demo.global.response.BaseResponse;
 import umc.cockple.demo.global.response.code.status.CommonSuccessCode;
 import umc.cockple.demo.global.security.utils.SecurityUtil;
