+++ src/main/java/umc/cockple/demo/domain/chat/service/command/ChatCommandServiceImpl.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.chat.service;
+package umc.cockple.demo.domain.chat.service.command;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
